+++ src/similarity.js
@@ -1,0 +1,9 @@
+function cosineSimilarity(vec1, vec2) {
+    const dot = vec1.reduce((sum, v, i) => sum + v * vec2[i], 0);
+    const norm1 = Math.sqrt(vec1.reduce((sum, v) => sum + v * v, 0));
+    const norm2 = Math.sqrt(vec2.reduce((sum, v) => sum + v * v, 0));
+    if (!norm1 || !norm2) return 0;
+    return dot / (norm1 * norm2);
+}
+
+module.exports = { cosineSimilarity };
